AdmissionType API INSERT,Update List,Delete

+++ Libraries/HIMS.Data/DTO/IPPatient/ItemNameBatchPOPIPPresReturnDto.cs
@@ -12,7 +12,7 @@ namespace HIMS.Data.DTO.IPPatient
         public string? BatchNo { get; set; }
         public DateTime? BatchExpDate { get; set; }
         public decimal? UnitMRP { get; set; }
-        public decimal? VatPer { get; set; }
+        public double? VatPer { get; set; }
         public double? Qty { get; set; }
         public decimal? LandedPrice { get; set; }
         public decimal? PurRateWf { get; set; }
